app/src/main/java/com/oracle/javafx/scenebuilder/app/util/TextInputControlPaster.java
/*
 * Copyright (c) 2022, Gluon and/or its affiliates.
 * All rights reserved. Use is subject to license terms.
 *
 * This file is available and licensed under the following license:
 *
 * Redistribution and use in source and binary forms, with or without
 * modification, are permitted provided that the following conditions
 * are met:
 *
 *  - Redistributions of source code must retain the above copyright
 *    notice, this list of conditions and the following disclaimer.
 *  - Redistributions in binary form must reproduce the above copyright
 *    notice, this list of conditions and the following disclaimer in
 *    the documentation and/or other materials provided with the distribution.
 *  - Neither the name of Oracle Corporation and Gluon nor the names of its
 *    contributors may be used to endorse or promote products derived
 *    from this software without specific prior written permission.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
 * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
 * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
 * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
 * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
 * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
 * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
 * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */

package com.oracle.javafx.scenebuilder.app.util;

import java.util.Objects;

import javafx.scene.control.TextInputControl;

public class TextInputControlPaster {
    private final TextInputControl tic;
    
    /**
     * This class provides a workaround for the issue in JavaFX
     * which prevents Scene Builder from pasting text values properly
     * in TextInputControl elements on MacOS. If attempted, clip board
     * contents is always pasted twice.
     * <p>
     * Given the way content pasted in Scene Builder is implemented now,
     * calling {@code textInputControl.paste()} method is not needed on MacOS.
     * 
     * @see https://bugs.openjdk.java.net/browse/JDK-8205915
     */
    public TextInputControlPaster(TextInputControl tic) {
        this.tic = Objects.requireNonNull(tic);
    }
    
    /**
     * @param isNotMacOS true when the current platform is not MacOS.
     */
    public void pasteConditionally(boolean isNotMacOS) {
        if (isNotMacOS) {
            tic.paste();
        }
    }
}
